refactor(tests): split parity tests into a kernel-agnostic package
tests/test_cpp_parity.py mixed three unrelated things: backend-dispatch
machinery, self-tests of the parity harness, and the parity/correctness
tests of the cardinal B-spline kernel. Both the dispatch tests and the
harness self-tests are kernel-agnostic, and three more kernels are about
to be ported through the same harness, so keeping them inside one
kernel's file meant copying two thirds of it four times.

Introduces tests/parity/ with conftest.py (the two session fixtures,
now shared across the package), test_dispatch.py (the backend/ISA
selection and use_backend's threading semantics), and
test_harness_selftest.py (the harness's own self-tests that do not
reach for a B-spline-specific point set). test_cpp_parity.py becomes
tests/parity/test_basis_cardinal_bspline.py, keeping everything
specific to that kernel, including the probes that drive the harness
through the cardinal B-spline's own point sets and so stayed kernel-
specific despite living next to the other harness self-tests.

Updates every citation of the old path across the C++ sources, the
build, and CI to point at the file or directory that now holds the
thing being cited.
