+++ tests/parity/__init__.py
@@ -1,0 +1,5 @@
+"""Kernel-agnostic C++ parity tests for PaNTr.
+
+Backend-dispatch machinery, the parity harness's own self-tests, and one module
+per ported kernel.
+"""
